Split NativeFunctions.cpp into functional components. (#4325)
